+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+Release v1.44.179 (2023-01-12)
+===
+
+### Service Client Updates
+* `service/cleanrooms`: Adds new service
+* `service/lambda`: Updates service API and documentation
+  * Add support for MaximumConcurrency parameter for SQS event source. Customers can now limit the maximum concurrent invocations for their SQS Event Source Mapping.
+* `service/logs`: Updates service API and documentation
+  * Bug fix: logGroupName is now not a required field in GetLogEvents, FilterLogEvents, GetLogGroupFields, and DescribeLogStreams APIs as logGroupIdentifier can be provided instead
+* `service/mediaconvert`: Updates service API and documentation
+  * The AWS Elemental MediaConvert SDK has added support for compact DASH manifest generation, audio normalization using TruePeak measurements, and the ability to clip the sample range in the color corrector.
+* `service/secretsmanager`: Updates service documentation and examples
+  * Update documentation for new ListSecrets and DescribeSecret parameters
+
 Release v1.44.178 (2023-01-11)
 ===